Extract query hint helpers partial
